refactor: Split UdfLoader into multiple files
