+++ ksql-engine/src/main/java/io/confluent/ksql/function/UdafLoader.java
@@ -1,0 +1,126 @@
+/*
+ * Copyright 2019 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.ksql.function;
+
+import io.confluent.ksql.function.udaf.UdafDescription;
+import io.confluent.ksql.function.udaf.UdafFactory;
+import io.confluent.ksql.function.udf.UdfMetadata;
+import io.confluent.ksql.name.FunctionName;
+import io.confluent.ksql.schema.ksql.SqlTypeParser;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import org.apache.kafka.common.metrics.Metrics;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+class UdafLoader {
+
+  private static final Logger LOGGER = LoggerFactory.getLogger(UdafLoader.class);
+
+  private final MutableFunctionRegistry functionRegistry;
+  private final Optional<Metrics> metrics;
+  private final SqlTypeParser typeParser;
+
+  UdafLoader(
+      final MutableFunctionRegistry functionRegistry,
+      final Optional<Metrics> metrics,
+      final SqlTypeParser typeParser
+  ) {
+    this.functionRegistry = functionRegistry;
+    this.metrics = metrics;
+    this.typeParser = typeParser;
+  }
+
+  void loadUdafFromClass(final Class<?> theClass, final String path) {
+    final UdafDescription udafAnnotation = theClass.getAnnotation(UdafDescription.class);
+    final List<UdafFactoryInvoker> argCreators
+        = Arrays.stream(theClass.getMethods())
+        .filter(method -> method.getAnnotation(UdafFactory.class) != null)
+        .filter(method -> {
+          if (!Modifier.isStatic(method.getModifiers())) {
+            LOGGER.warn(
+                "Trying to create a UDAF from a non-static factory method. Udaf factory"
+                    + " methods must be static. class={}, method={}, name={}",
+                method.getDeclaringClass(),
+                method.getName(),
+                udafAnnotation.name()
+            );
+            return false;
+          }
+          return true;
+        })
+        .map(method -> {
+          final UdafFactory annotation = method.getAnnotation(UdafFactory.class);
+          try {
+            LOGGER.info(
+                "Adding UDAF name={} from path={} class={}",
+                udafAnnotation.name(),
+                path,
+                method.getDeclaringClass()
+            );
+            return Optional.of(createUdafFactoryInvoker(
+                method,
+                FunctionName.of(udafAnnotation.name()),
+                annotation.description(),
+                annotation.paramSchema(),
+                annotation.aggregateSchema(),
+                annotation.returnSchema()
+            ));
+          } catch (final Exception e) {
+            LOGGER.warn(
+                "Failed to create UDAF name={}, method={}, class={}, path={}",
+                udafAnnotation.name(),
+                method.getName(),
+                method.getDeclaringClass(),
+                path,
+                e
+            );
+          }
+          return Optional.<UdafFactoryInvoker>empty();
+        }).filter(Optional::isPresent)
+        .map(Optional::get)
+        .collect(Collectors.toList());
+
+    functionRegistry.addAggregateFunctionFactory(new UdafAggregateFunctionFactory(
+        new UdfMetadata(
+            udafAnnotation.name(),
+            udafAnnotation.description(),
+            udafAnnotation.author(),
+            udafAnnotation.version(),
+            path,
+            false
+        ),
+        argCreators
+    ));
+  }
+
+  UdafFactoryInvoker createUdafFactoryInvoker(
+      final Method method,
+      final FunctionName functionName,
+      final String description,
+      final String inputSchema,
+      final String aggregateSchema,
+      final String outputSchema
+  ) {
+    return new UdafFactoryInvoker(method, functionName, description, inputSchema,
+        aggregateSchema, outputSchema, typeParser, metrics
+    );
+  }
+}
